Move DefaultBinarySerializer tests to its own file
